video: fix incorrect framerate selection
- give libav a rational<double> instead of rational<unsigned> for framerate
- add libav patch for osx to correct this issue

Tuleap: #231

+++ src/media/media_decoder.h
@@ -87,7 +87,7 @@ class MediaDecoder {
 
         int getWidth() const;
         int getHeight() const;
-        rational<unsigned> getFps() const;
+        rational<double> getFps() const;
         int getPixelFormat() const;
 
         void setOptions(const std::map<std::string, std::string>& options);
